Re-type ResistInfo.cs#ResistPercent from `int` to `decimal`

+++ Sanctuary.Census.ClientData/ClientDataModels/ResistInfo.cs
@@ -17,7 +17,7 @@ public partial record ResistInfo
 (
     uint ID,
     ushort ResistType,
-    int ResistPercent,
+    decimal ResistPercent,
     int ResistAmount,
     decimal MultiplierWhenHeadshot,
     uint ResistIcon,
